自定义输出宏

+++ TJTools/TJTools/TJNSString/TJ_RSA.m
@@ -302,7 +302,7 @@ static NSData *base64_decode(NSString *str){
         status = SecKeyDecrypt(keyRef, kSecPaddingNone, srcbuf + idx, data_len, outbuf, &outlen);
         
         if (status != 0) {
-            NSLog(@"SecKeyEncrypt fail. Error Code: %d", status);
+            NSLog(@"SecKeyEncrypt fail. Error Code: %d", (int)status);
             ret = nil;
             break;
         }else {
